Fix crash on 32-bit processor devices.

+++ Swamp/Messages/Error/ErrorSwampMessage.swift
@@ -11,14 +11,14 @@ import Foundation
 /// [ERROR, requestType|number, requestId|number, details|dict, error|string, args|array?, kwargs|dict?]
 class ErrorSwampMessage: SwampMessage {
     let requestType: SwampMessages
-    let requestId: Int
+    let requestId: UInt64
     let details: [String: Any]
     let error: String
     
     let args: [Any]?
     let kwargs: [String: Any]?
     
-    init(requestType: SwampMessages, requestId: Int, details: [String: Any], error: String, args: [Any]?=nil, kwargs: [String: Any]?=nil) {
+    init(requestType: SwampMessages, requestId: UInt64, details: [String: Any], error: String, args: [Any]?=nil, kwargs: [String: Any]?=nil) {
         self.requestType = requestType
         self.requestId = requestId
         self.details = details
@@ -31,7 +31,7 @@ class ErrorSwampMessage: SwampMessage {
     
     required init(payload: [Any]) {
         self.requestType = SwampMessages(rawValue: payload[0] as! Int)!
-        self.requestId = payload[1] as! Int
+        self.requestId = payload[1] as! UInt64
         self.details = payload[2] as! [String: Any]
         self.error = payload[3] as! String
