+++ Swamp/Messages/PubSub/Publisher/PublishedSwampMessage.swift
@@ -11,18 +11,18 @@ import Foundation
 /// [PUBLISHED, requestId|number, options|dict, topic|String, args|list?, kwargs|dict?]
 class PublishedSwampMessage: SwampMessage {
     
-    let requestId: Int
-    let publication: Int
+    let requestId: UInt64
+    let publication: UInt64
     
-    init(requestId: Int, publication: Int) {
+    init(requestId: UInt64, publication: UInt64) {
         self.requestId = requestId
         self.publication = publication
     }
     
     // MARK: SwampMessage protocol
     required init(payload: [Any]) {
-        self.requestId = payload[0] as! Int
-        self.publication = payload[1] as! Int
+        self.requestId = payload[0] as! UInt64
+        self.publication = payload[1] as! UInt64
     }
     
     func marshal() -> [Any] {
